Refactor:changed name of MarorQue.tsx and moved to ~/components/Icons
