+++ app/components/Icons/MayorQue.tsx
